requirements.txt
flask
stripe 
python-dotenv